refactor(server): /api/repositories + projects-health + health 라우트 분할 (Plan C #C9 9단계)
- src/server/health-checks.ts 신규 (111줄): 4개 헬퍼 추출
  (checkGitRemoteAccess/checkLocalPath/checkDiskSpace/checkDependencies)
  세 라우트가 공통으로 사용하던 인라인 헬퍼 단일화
- src/server/routes/health.ts 신규 (251줄): 3개 라우트 캡슐화
  - GET /api/repositories (저장소 목록 + 헬스 + 통계)
  - GET /api/projects/health (전체 프로젝트 헬스)
  - GET /api/health (단일 프로젝트 헬스)
- runProjectChecks/deriveOverallStatus 헬퍼로 3 라우트 간 중복 정리
- safeError 적용
- dashboard-api.ts unused import 정리 (existsSync/statSync/HealthCheckResponse/getProjectSummary)
- dashboard-api.ts 983 → 661줄 (-322) — **이번 단계 누계 최대 감축**

남은 도메인: skip-events/sse/auth/new-issue
